Layout Finished.
Depending on what page you are on, the navigation bar is highlighted.
The red color was chosen as a complement to this green according to
some online color theory website.
Moved the navigation to the top of the page to follow Android Design
Patterns.

Disclaimer I have not taken screenshots, but I did change the layout
based on Nexus One, so hopefully this time it is not shrunk like how
Daniel has reported over this weekend.